Add UserTable And Roles And Fixed Select Bugs In Product View

+++ E_CommerceStore_Udemey.WEB/Areas/Admin/Controllers/ProductController.cs
@@ -77,27 +77,22 @@ namespace E_CommerceStore_Udemey.WEB.Controllers
             return View(product);
         }
 
-        //[HttpPost]
-        //[ValidateAntiForgeryToken]
-        //public async Task<IActionResult> Edit(UpdateProductDto  dto)
-        //{
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> Edit(UpdateProductDto dto)
+        {
 
-        //    //// validMessage in Sammerry just appear
-        //    //if (dto.Name == dto.DisplayOrder.ToString())
-        //    //{
-        //    //    ModelState.AddModelError("CustomErorr", "The DisplayOrder cannot exactly math the Name.");
-        //    //}
-        //    if (ModelState.IsValid)
-        //    {
-        //        ViewData["category"] = new SelectList(await _categoryService.GetCategoryName(), "Id", "Name");
-        //        ViewData["coverType"] = new SelectList(await _coverTypeService.GetCoverTypeName(), "Id", "CoverName");
+            if (ModelState.IsValid)
+            {
+                ViewData["category"] = new SelectList(await _categoryService.GetCategoryName(), "Id", "Name");
+                ViewData["coverType"] = new SelectList(await _coverTypeService.GetCoverTypeName(), "Id", "CoverName");
 
-        //        await _productService.Update(dto);
-        //        TempData["success"] = "Edit Product Successfully";
-        //        return RedirectToAction("Index");
-        //    }
-        //    else { return View(dto); }
-        //}
+                await _productService.Update(dto);
+                TempData["success"] = "Edit Product Successfully";
+                return RedirectToAction("Index");
+            }
+            else { return View(dto); }
+        }
 
 
 
@@ -105,10 +100,10 @@ namespace E_CommerceStore_Udemey.WEB.Controllers
         [HttpDelete]
         public async Task<IActionResult> Delete(int id)
         {
-            //if (id == null)
-            //{
-            //    return Json(new { success = false, message = "Error while deleting" });
-            //}
+            if (id == null)
+            {
+                return Json(new { success = false, message = "Error while deleting" });
+            }
             await _productService.Delete(id);
             return Json(new { success = true, message = "Delete Successful" });
